PIM-4252: move the field splitter
