+++ src/Pim/Component/Connector/ArrayConverter/Flat/Product/FieldSplitter.php
@@ -1,8 +1,6 @@
 <?php
 
-namespace Pim\Component\Connector\ArrayConverter\Flat\Product\Splitter;
-
-use Pim\Component\Connector\ArrayConverter\Flat\Product\AttributeColumnInfoExtractor;
+namespace Pim\Component\Connector\ArrayConverter\Flat\Product;
 
 /**
  * Split fields
@@ -10,8 +8,6 @@ use Pim\Component\Connector\ArrayConverter\Flat\Product\AttributeColumnInfoExtra
  * @author    Olivier Soulet <olivier.soulet@akeneo.com>
  * @copyright 2015 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
- *
- * TODO : two concerns
  */
 class FieldSplitter
 {
